move files around, add local stubs for MESH linker
